A tag-based metadata ontology
The record-schema link is removed. The ODP tagging framework provides a way to define a unified, standards-independent ontology for describing digital objects.

ISO 19115, DataCite, Darwin Core, etc, can be used to validate incoming package metadata from data providers. Thus, we can publish an upload API based on internationally-recognised metadata standards. Similarly, we can publish catalog API output endpoints conforming to recognised standards. There is, however, no advantage in maintaining metadata records internally as JSON blobs.

JSON is a transport format which we are currently using as a storage format. It's difficult to make ad-hoc queries for deeply nested values, and it's practically impossible to track changes to individual elements, to mention a couple of challenges with this approach. Using tags to store metadata values, we can standardise our internal model and vocabularies; we get per-client and per-user tracking of individual field changes; and we can systematically model, normalize and reuse components of complex elements - affiliations, contact info, and so on.